Refactor Advent of Code project
Create an AdventSupport framework to facilitate code sharing between
solutions. Also move the individual solutions under a common project.
Additionally, create a single main Swift file that can enumerate the
project solutions in the application and execute one based on the
supplied command-line arguments. This was done to support a separate
Advent of Code 2016 solutions project.