Migrate to Swift 4
Swift 4 is mostly but not entirely source compatible with Swift 3. The
biggest change for Swift 4 was the breakage introduced by the improved
NSNumber bridging. This required my day 12 solution to peak at the
original C type to filter out boolean values from the JSON.
Unfortunately, JSONDecoder is not flexible enough to take arbitrary
JSON. It’s designed to consume structured JSON.